Extract backward compatibility from aftership-functions
